+++ crates/cheatnet/src/cheatcodes.rs
@@ -16,6 +16,7 @@ pub mod mock_call;
 pub mod prank;
 pub mod precalculate_address;
 pub mod roll;
+pub mod spy_events;
 pub mod warp;
 
 // All errors that can be thrown from the hint executor have to be added here,