+++ crates/forge/tests/integration/mod.rs
@@ -10,5 +10,6 @@ mod precalculate_address;
 mod pure_cairo;
 mod roll;
 mod should_panic;
+mod spy_events;
 mod syscalls;
 mod warp;